@@ -1,6 +1,6 @@
 {
   "name": "jsonwebtoken",
-  "version": "8.5.1",
+  "version": "9.0.2",
   "description": "JSON Web Token implementation (symmetric and asymmetric)",
   "main": "index.js",
   "nyc": {
@@ -45,7 +45,7 @@
     "lodash.isstring": "^4.0.1",
     "lodash.once": "^4.0.0",
     "ms": "^2.1.1",
-    "semver": "^5.6.0"
+    "semver": "^7.5.4"
   },
   "devDependencies": {
     "atob": "^2.1.2",
@@ -59,8 +59,8 @@
     "sinon": "^6.0.0"
   },
   "engines": {
-    "npm": ">=1.4.28",
-    "node": ">=4"
+    "npm": ">=6",
+    "node": ">=12"
   },
   "files": [
     "lib",
