@@ -1,4 +1,4 @@
-import './index.css'
+import './index.css';
 import React, { Component } from 'react';
 import { Navigate } from 'react-router-dom';
 
@@ -15,17 +15,15 @@ class LoginPage extends Component {
   }
 
   handleInputChange = (event) => {
-    const { name, value, type, checked } = event.target;
-    this.setState({
-      [name]: type === 'checkbox' ? checked : value,
-    });
+    const { name, value } = event.target;
+    this.setState({ [name]: value });
   };
 
   handleLogin = async (event) => {
     event.preventDefault();
     const { email, password } = this.state;
     const { onLoginButton } = this.props;
-
+  
     try {
       const response = await fetch('http://localhost:5000/login', {
         method: 'POST',
@@ -34,19 +32,23 @@ class LoginPage extends Component {
         },
         body: JSON.stringify({ email, password }),
       });
-
-      const data = await response.json();
-      if (response.ok) {
+  
+      if (response.status === 200) {
+        const data = await response.json();
         console.log('Login successful:', data);
         this.setState({ redirect: '/' });
-        onLoginButton({ name: data.name, email: email, password: password });
+        onLoginButton({ name: data.name, email, password });
       } else {
-        console.error('Login failed:', data);
+        const errorText = await response.text();
+        console.error('Login failed:', errorText);
+        this.setState({ error: errorText });
       }
     } catch (error) {
-      console.error('Error logging in:', error);
+      console.error('Error logging in:', error.message || error);
+      this.setState({ error: 'An unexpected error occurred.' });
     }
   };
+  
 
   handleRegister = async (event) => {
     event.preventDefault();
@@ -62,14 +64,13 @@ class LoginPage extends Component {
         body: JSON.stringify({ name, email, password }),
       });
 
-      if (response.ok) {
+      if (response.status === 200) {
         console.log('Registration successful');
         this.setState({ redirect: '/' });
         localStorage.setItem('userDetailsMovie', JSON.stringify({ name, email, password }));
         onRegisterButton({ name, email, password });
-        return (<Navigate to={'/'} />)
       } else {
-        console.error('Registration failed');
+        console.error('Registration failed:', await response.text());
       }
     } catch (error) {
       console.error('Error registering:', error);
@@ -85,7 +86,11 @@ class LoginPage extends Component {
 
     return (
       <div className='elements-container'>
-        <img src="https://imgs.search.brave.com/PBObDQhElLLghbtGfzIbbeZ_G5wGbf2rCuVZ7CMUum4/rs:fit:860:0:0:0/g:ce/aHR0cHM6Ly9jZG4u/aGFzaG5vZGUuY29t/L3Jlcy9oYXNobm9k/ZS9pbWFnZS91cGxv/YWQvdjE2NDc0OTA2/MTk5NjUvUDFkc05n/ai1mMS5wbmc" alt="react" className="img-style"/>
+        <img
+          src="https://imgs.search.brave.com/PBObDQhElLLghbtGfzIbbeZ_G5wGbf2rCuVZ7CMUum4/rs:fit:860:0:0:0/g:ce/aHR0cHM6Ly9jZG4u/aGFzaG5vZGUuY29t/L3Jlcy9oYXNobm9k/ZS9pbWFnZS91cGxv/YWQvdjE2NDc0OTA2/MTk5NjUvUDFkc05n/ai1mMS5wbmc"
+          alt="react"
+          className="img-style"
+        />
         {isLoginForm ? (
           <form onSubmit={this.handleLogin} className="form">
             <h2 className='login-header'>Login</h2>
@@ -106,7 +111,7 @@ class LoginPage extends Component {
               className='input'
             />
             <button type="submit" className="button">Login</button>
-            <p>Don't have an account? <button type="button" className="txt-decoration" onClick={() => this.setState({ isLoginForm: false })} style={{ background: 'none', border: 'none', color: 'blue', textDecoration: 'none', cursor: 'pointer' }}>Register</button></p>
+            <p>Don't have an account? <button type="button" className="txt-decoration" onClick={() => this.setState({ isLoginForm: false })} style={{ background: 'none', border: 'none', color: 'blue', cursor: 'pointer' }}>Register</button></p>
           </form>
         ) : (
           <form onSubmit={this.handleRegister} className="form">
@@ -136,7 +141,7 @@ class LoginPage extends Component {
               className='input'
             />
             <button type="submit" className='button'>Register</button>
-            <p>Already have an account? <button type="button" className="text-btn" onClick={() => this.setState({ isLoginForm: true })} style={{ background: 'none', border: 'none', color: 'blue', textDecoration: 'none', cursor: 'pointer' }}>Login</button></p>
+            <p>Already have an account? <button type="button" className="text-btn" onClick={() => this.setState({ isLoginForm: true })} style={{ background: 'none', border: 'none', color: 'blue', cursor: 'pointer' }}>Login</button></p>
           </form>
         )}
       </div>
